优化zookeeper是否存在节点数据的检查方式

+++ src/Surging.Core/Surging.Core.Zookeeper/ZooKeeperMqttServiceRouteManager.cs
@@ -69,7 +69,7 @@ namespace Surging.Core.Zookeeper
                 {
                     var nodePath = "/" + string.Join("/", childrens);
 
-                    if (await zooKeeper.Item2.existsAsync(nodePath) != null)
+                    if (await zooKeeper.Item2.Exists(nodePath))
                     {
                         var result = await zooKeeper.Item2.getChildrenAsync(nodePath);
                         if (result?.Children != null)
@@ -79,7 +79,11 @@ namespace Surging.Core.Zookeeper
                                 var childPath = $"{nodePath}/{child}";
                                 if (_logger.IsEnabled(LogLevel.Debug))
                                     _logger.LogDebug($"准备删除：{childPath}。");
-                                await zooKeeper.Item2.deleteAsync(childPath);
+                                if (await zooKeeper.Item2.Exists(childPath)) 
+                                {
+                                    await zooKeeper.Item2.deleteAsync(childPath);
+                                }
+                               
                             }
                         }
                         if (_logger.IsEnabled(LogLevel.Debug))
@@ -116,7 +120,7 @@ namespace Surging.Core.Zookeeper
                 {
                     var nodePath = $"{path}{serviceRoute.MqttDescriptor.Topic}";
                     var nodeData = _serializer.Serialize(serviceRoute);
-                    if (await zooKeeper.Item2.existsAsync(nodePath) == null)
+                    if (!await zooKeeper.Item2.Exists(nodePath))
                     {
                         if (_logger.IsEnabled(LogLevel.Debug))
                             _logger.LogDebug($"节点：{nodePath}不存在将进行创建。");
@@ -204,7 +208,11 @@ namespace Surging.Core.Zookeeper
                         if (addresses.Contains(hostAddr))
                         {
                             var nodePath = $"{path}{deletedRouteTopic}";
-                            await zooKeeper.Item2.deleteAsync(nodePath);
+                            if (await zooKeeper.Item2.Exists(nodePath)) 
+                            {
+                                await zooKeeper.Item2.deleteAsync(nodePath);
+                            }
+                            
                         }
                     }
                 }
@@ -214,7 +222,7 @@ namespace Surging.Core.Zookeeper
         private async Task CreateSubdirectory((ManualResetEvent, ZooKeeper) zooKeeper, string path)
         {
             zooKeeper.Item1.WaitOne();
-            if (await zooKeeper.Item2.existsAsync(path) != null)
+            if (await zooKeeper.Item2.Exists(path))
                 return;
 
             if (_logger.IsEnabled(LogLevel.Information))
@@ -226,7 +234,7 @@ namespace Surging.Core.Zookeeper
             foreach (var children in childrens)
             {
                 nodePath += children;
-                if (await zooKeeper.Item2.existsAsync(nodePath) == null)
+                if (!await zooKeeper.Item2.Exists(nodePath))
                 {
                     await zooKeeper.Item2.createAsync(nodePath, null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
                 }
@@ -253,7 +261,7 @@ namespace Surging.Core.Zookeeper
             var zooKeeper = await GetZooKeeper();
             var watcher = new NodeMonitorWatcher(GetZooKeeper, path,
                  async (oldData, newData) => await NodeChange(oldData, newData));
-            if (await zooKeeper.Item2.existsAsync(path) != null)
+            if (await zooKeeper.Item2.Exists(path))
             {
                 var data = (await zooKeeper.Item2.getDataAsync(path, watcher)).Data;
                 watcher.SetCurrentData(data);
@@ -291,7 +299,7 @@ namespace Surging.Core.Zookeeper
             zooKeeper.Item1.WaitOne(); 
             var watcher = new ChildrenMonitorWatcher(GetZooKeeper, _configInfo.MqttRoutePath,
              async (oldChildrens, newChildrens) => await ChildrenChange(oldChildrens, newChildrens));
-            if (await zooKeeper.Item2.existsAsync(_configInfo.MqttRoutePath, watcher) != null)
+            if (await zooKeeper.Item2.Exists(_configInfo.MqttRoutePath, watcher))
             {
                 var result = await zooKeeper.Item2.getChildrenAsync(_configInfo.MqttRoutePath, watcher);
                 var childrens = result.Children.ToArray();
