+++ src/Surging.Core/Surging.Core.Zookeeper/ZooKeeperServiceSubscribeManager.cs
@@ -65,7 +65,7 @@ namespace Surging.Core.Zookeeper
                 {
                     var nodePath = "/" + string.Join("/", childrens);
 
-                    if (await zooKeeper.Item2.existsAsync(nodePath) != null)
+                    if (await zooKeeper.Item2.Exists(nodePath))
                     {
                         var result = await zooKeeper.Item2.getChildrenAsync(nodePath);
                         if (result?.Children != null)
@@ -75,7 +75,11 @@ namespace Surging.Core.Zookeeper
                                 var childPath = $"{nodePath}/{child}";
                                 if (_logger.IsEnabled(LogLevel.Debug))
                                     _logger.LogDebug($"准备删除：{childPath}。");
-                                await zooKeeper.Item2.deleteAsync(childPath);
+                                if (await zooKeeper.Item2.Exists(childPath)) 
+                                {
+                                    await zooKeeper.Item2.deleteAsync(childPath);
+                                }
+                                
                             }
                         }
                         if (_logger.IsEnabled(LogLevel.Debug))
@@ -126,7 +130,7 @@ namespace Surging.Core.Zookeeper
                 {
                     var nodePath = $"{path}{serviceSubscriber.ServiceDescriptor.Id}";
                     var nodeData = _serializer.Serialize(serviceSubscriber);
-                    if (await zooKeeper.Item2.existsAsync(nodePath) == null)
+                    if (!await zooKeeper.Item2.Exists(nodePath))
                     {
                         if (_logger.IsEnabled(LogLevel.Debug))
                             _logger.LogDebug($"节点：{nodePath}不存在将进行创建。");
@@ -167,7 +171,7 @@ namespace Surging.Core.Zookeeper
         private async Task CreateSubdirectory((ManualResetEvent, ZooKeeper) zooKeeper, string path)
         {
             zooKeeper.Item1.WaitOne();
-            if (await zooKeeper.Item2.existsAsync(path) != null)
+            if (await zooKeeper.Item2.Exists(path))
                 return;
 
             if (_logger.IsEnabled(LogLevel.Information))
@@ -179,7 +183,7 @@ namespace Surging.Core.Zookeeper
             foreach (var children in childrens)
             {
                 nodePath += children;
-                if (await zooKeeper.Item2.existsAsync(nodePath) == null)
+                if (!await zooKeeper.Item2.Exists(nodePath) )
                 {
                     await zooKeeper.Item2.createAsync(nodePath, null, ZooDefs.Ids.OPEN_ACL_UNSAFE, CreateMode.PERSISTENT);
                 }
@@ -204,7 +208,7 @@ namespace Surging.Core.Zookeeper
             ServiceSubscriber result = null;
 
             var zooKeeper = await GetZooKeeper();
-            if (await zooKeeper.Item2.existsAsync(path) != null)
+            if (await zooKeeper.Item2.Exists(path))
             {
                 var data = (await zooKeeper.Item2.getDataAsync(path)).Data;
                 result = await GetSubscriber(data);
@@ -241,7 +245,7 @@ namespace Surging.Core.Zookeeper
             var zooKeeper = await GetZooKeeper();
             zooKeeper.Item1.WaitOne(); 
 
-            if (await zooKeeper.Item2.existsAsync(_configInfo.SubscriberPath) != null)
+            if (await zooKeeper.Item2.Exists(_configInfo.SubscriberPath))
             {
                 var result = await zooKeeper.Item2.getChildrenAsync(_configInfo.SubscriberPath);
                 var childrens = result.Children.ToArray();
